client: check graphics executable before telling GUI about it
