+++ client/client_state.cpp
@@ -15,17 +15,6 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with BOINC.  If not, see <http://www.gnu.org/licenses/>.
 
-#ifdef __APPLE__
-#include <sys/types.h>
-#include <sys/sysctl.h>
-#include <mach-o/loader.h>
-#include <mach-o/fat.h>
-#include <mach/machine.h>
-#include <libkern/OSByteOrder.h>
-
-extern int compareOSVersionTo(int toMajor, int toMinor);
-#endif
-
 #ifdef _WIN32
 #include "boinc_win.h"
 #else
@@ -1325,17 +1314,7 @@ int CLIENT_STATE::link_app_version(PROJECT* p, APP_VERSION* avp) {
         }
 
         if (!strcmp(file_ref.open_name, GRAPHICS_APP_FILENAME)) {
-            char relpath[MAXPATHLEN], path[MAXPATHLEN];
-            get_pathname(fip, relpath, sizeof(relpath));
-            relative_to_absolute(relpath, path);
-#ifdef __APPLE__
-            if (can_run_on_this_CPU(path))
-            
-#endif
-            {
-                safe_strcpy(avp->graphics_exec_path, path);
-                safe_strcpy(avp->graphics_exec_file, fip->name);
-            }
+            avp->graphics_exec_fip = fip;
         }
 
         // any file associated with an app version must be signed
@@ -2373,125 +2352,3 @@ bool CLIENT_STATE::abort_sequence_done() {
 }
 
 #endif
-
-#ifdef __APPLE__
-
-union headeru {
-    fat_header fat;
-    mach_header mach;
-};
-// Get the architecture of this computer's CPU: x86_64 or arm64.
-// Read the executable file's mach-o headers to determine the 
-// architecture(s) of its code.
-// Returns 1 if application can run natively on this computer,
-// else returns 0.
-//
-// ToDo: determine whether x86_64 graphics apps emulated on arm64 Macs 
-// properly run under Rosetta 2. Note: years ago, PowerPC apps emulated 
-// by Rosetta on i386 Macs crashed when running graphics.
-//
-
-int CLIENT_STATE::can_run_on_this_CPU(char* exec_path) {
-    FILE *f;
-    int retval = 0;
-    
-    headeru myHeader;
-    fat_arch fatHeader;
-    
-    static bool x86_64_CPU = false;
-    static bool arm64_cpu = false;
-    static bool need_CPU_architecture = true;
-    uint32_t n, i, len;
-    uint32_t theMagic;
-    integer_t file_architecture;
-    
-    if (need_CPU_architecture) {
-        // Determine the architecture of the CPU we are running on
-        // ToDo: adjust this code accordingly.
-        uint32_t cputype = 0;
-        size_t size = sizeof (cputype);
-        int res = sysctlbyname ("hw.cputype", &cputype, &size, NULL, 0);
-        if (res) return false;  // Should never happen
-        // Since we require MacOS >= 10.7, the CPU must be x86_64 or arm64 
-        x86_64_CPU = ((cputype &0xff) == CPU_TYPE_X86);
-        arm64_cpu = ((cputype &0xff) == CPU_TYPE_ARM);
-
-        need_CPU_architecture = false;
-    }
-    
-    f = boinc_fopen(exec_path, "rb");
-    if (!f) {
-        return retval;          // Should never happen
-    }
-    
-    myHeader.fat.magic = 0;
-    myHeader.fat.nfat_arch = 0;
-    
-    fread(&myHeader, 1, sizeof(fat_header), f);
-    theMagic = myHeader.mach.magic;
-    switch (theMagic) {
-    case MH_CIGAM:
-    case MH_MAGIC:
-    case MH_MAGIC_64:
-    case MH_CIGAM_64:
-       file_architecture = myHeader.mach.cputype;
-        if ((theMagic == MH_CIGAM) || (theMagic == MH_CIGAM_64)) {
-            file_architecture = OSSwapInt32(file_architecture);
-        }
-        if (x86_64_CPU && (file_architecture == CPU_TYPE_I386)) {
-            // Single-architecture i386 file on x86_64 CPU
-            if (compareOSVersionTo(10, 15) < 0) {   // OS >= 10.15 are 64-bit only
-                retval = 1;
-            }
-        } else
-        if (x86_64_CPU && (file_architecture == CPU_TYPE_X86_64)) {
-            retval = 1; // Single-architecture x86_64 file on x86_64 CPU
-        } else
-        if (arm64_cpu && (file_architecture == CPU_TYPE_ARM64)) {
-            retval = 1; // Single-architecture arm64 file on arm64 CPU
-        } else
-        if (arm64_cpu && (file_architecture == CPU_TYPE_X86_64)) {
-            retval = 1; // Single-architecture x86_64 file emulated on arm64 CPU
-            // ToDo: determine whether emulated graphics apps work properly
-        }
-        break;
-    case FAT_MAGIC:
-    case FAT_CIGAM:
-        n = myHeader.fat.nfat_arch;
-        if (theMagic == FAT_CIGAM) {
-            n = OSSwapInt32(myHeader.fat.nfat_arch);
-        }
-           // Multiple architecture (fat) file
-        for (i=0; i<n; i++) {
-            len = fread(&fatHeader, 1, sizeof(fat_arch), f);
-            if (len < sizeof(fat_arch)) {
-                break;          // Should never happen
-            }
-            file_architecture = fatHeader.cputype;
-            if (theMagic == FAT_CIGAM) {
-                file_architecture = OSSwapInt32(file_architecture);
-            }
-
-            if (x86_64_CPU && (file_architecture == CPU_TYPE_X86_64)) {
-                retval = 1; // file with x86_64 architecture on x86_64 CPU
-                            break;
-            } else
-            if (arm64_cpu && (file_architecture == CPU_TYPE_ARM64)) {
-                retval = 1; // file with arm64 architecture on arm64 CPU
-                break;
-            } else
-            if (arm64_cpu && (file_architecture == CPU_TYPE_X86_64)) {
-                retval = 1; // file with x86_64 architecture emulated on arm64 CPU
-                // ToDo: determine whether emulated graphics apps work properly
-            }
-        }
-        break;
-    default:
-        break;
-    }
-
-    fclose (f);
-    return retval;
-}
-#endif
-
